bug fixed
bug 88

Idea how this can be solved:

1) If a new (so far invisible) individual is created while creating a new
property member, a dialog can be opened by the NTK (preferences) to ask the
user if he want to create a class assertion axiom to owl:Thing (see deleting
last class assertion of an individual)

and

2) The IndivialView could also display unasserted (so far invisible)
individuals, if owl:Thing is selected. (Maybe optional, switch in the
IndividualView)

and

3) Additionally a feature can be addded to look for unasserted individuals and
insert assertions to owl:Thing or an specific class. This could be added in the
context menu for ontologies or projects in the Ontology Navigator and/or in the
menu tools or the menu refactoring in the menubar.


1, 2 and 3 are implemented