bug fixed bug #527
problem fixed:
when hitting CTRL-SHIFT-N on a class, the individual view show as blinking
cursor for entering the name of the individual. Upon RETURN nothing happens.

Somehow I managed to create an individual. Strangely enough this is displayed
in the Individual View all the time, even if I select another class or
ontology.

I clicked the "show individuals of subclasses" button a few time before i
discovered this bug.