bugs fixed
bug #74
bug #537
bug #543

bug reopend

bug #544

loading process has been changed:
not is possible to load ontologies which import other ontologies which does not exist.
- user is informed with a message
- this ontologies are marked 
   * with an different icon in the Ontology Navigator and 
   * red in the import area of the importing ontologyr
- it is possible to add this ontologies manually
   * from Ontology Navigator and
   * in the imports area of the importing ontology

sadly there is a problem with the latest version the OWL API(3.2.3 and the NTK, the import is not working anymore), so it was neccesary to switch back to the 3.2.0, but all the rest is prepared for the latest version, so after this bug is fixed, it can be switched to the latest release without any further problems.